+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2024_10_26_215202) do
+ActiveRecord::Schema[7.0].define(version: 2024_10_28_153442) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -47,6 +47,7 @@ ActiveRecord::Schema[7.0].define(version: 2024_10_26_215202) do
     t.datetime "updated_at", null: false
     t.string "event_avatar"
     t.bigint "user_id", default: 1
+    t.string "external_event_id"
     t.index ["user_id"], name: "index_events_on_user_id"
   end
